Add DB::transations to destroy methods | Add 'Delete with relations' option to user admin management

+++ app/Services/UserDeleteServices.php
@@ -9,15 +9,23 @@ use Illuminate\Support\Facades\Storage;
 
 final class UserDeleteServices
 {
-	public static function deleteUser(User $user): void
+	public static function deleteUser(User $user, bool $with_relations = false): void
 	{
-		DB::transaction(function () use ($user) {
+		DB::transaction(function () use ($user, $with_relations) {
 			$pfp = $user->avatar;
 
-			$user->games()->delete();
-			$user->reviews()->update(['user_id' => null]);
-			$user->discussions()->update(['user_id' => null]);
-			$user->comments()->update(['user_id' => null]);
+			if ($with_relations) {
+				$user->games()->delete();
+				$user->reviews()->delete();
+				$user->discussions()->delete();
+				$user->comments()->delete();
+			} else {
+				$user->games()->update(['user_id' => null]);
+				$user->reviews()->update(['user_id' => null]);
+				$user->discussions()->update(['user_id' => null]);
+				$user->comments()->update(['user_id' => null]);
+			}
+
 			GameCreatorRequest::where('user_id', $user->id)->first()?->delete();
 			$user->delete();
